requirements.txt
numpy
PyQt5
PyQt5-stubs
PyQtWebEngine
Pillow
pyscreenshot
tensorflow
scipy
matplotlib
chess
pyautogui
